requirements.txt
appdirs==1.4.0
appJar==0.42
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
